[refactor] Rename common to ps-common-lib
