+++ non-atomic-action-interpreter-module/CMakeLists.txt
@@ -72,7 +72,7 @@ set(CMAKE_BUILD_RPATH_USE_ORIGIN TRUE)
 set(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
 
 find_package(sc-machine REQUIRED)
-find_package(common REQUIRED)
+find_package(ps-common-lib REQUIRED)
 
 add_subdirectory(non-atomic-action-interpreter-module)
