+++ non-atomic-action-interpreter-module/non-atomic-action-interpreter-module/interpreter/NonAtomicActionInterpreter.cpp
@@ -3,10 +3,10 @@
 #include <algorithm>
 
 #include <sc-agents-common/utils/IteratorUtils.hpp>
-#include <common/action_cancelled_exception.hpp>
-#include <common/utils/logic_utils.hpp>
-#include <common/utils/template_params_utils.hpp>
-#include <common/utils/macros.hpp>
+#include <ps-common-lib/action_cancelled_exception.hpp>
+#include <ps-common-lib/utils/logic_utils.hpp>
+#include <ps-common-lib/utils/template_params_utils.hpp>
+#include <ps-common-lib/utils/macros.hpp>
 
 #include "constants/NonAtomicActionInterpreterConstants.hpp"
 #include "keynodes/NonAtomicKeynodes.hpp"
