+++ non-atomic-action-interpreter-module/specification.scs
@@ -18,7 +18,7 @@ non_atomic_action_interpreter_agent_specification = [*
 
     => nrel_main_idtf: [Агент интерпретации неатомарных действий] (* <- lang_ru;; *);;
     => nrel_main_idtf: [Non-atomic action interpreter agent] (* <- lang_en;; *);;
-    => nrel_component_dependencies: ... (* <- empty_set;; *);; // TODO(NikitaZotov): Add sc-machine and common-lib dependencies
+    => nrel_component_dependencies: ... (* <- empty_set;; *);; // TODO(NikitaZotov): Add sc-machine and ps-common-lib-lib dependencies
 
     => nrel_component_address: [https://github.com/NikitaZotov/ostis-ps-lib/tree/feat/move_common_and_non_atomic_action_interpreter_module_to_new_build_system/non-atomic-action-interpreter-module] (* <- concept_github_url;; *);;
     => nrel_installation_method: ... (* <- concept_component_dynamically_installed_method;; *);; // TODO(NikitaZotov): Add installation method
