+++ ps-common-lib/cmake/install.cmake
@@ -1,6 +1,6 @@
 install(TARGETS 
     common-utils
-    EXPORT commonExport
+    EXPORT ps-common-lib-export
     LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
     ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
     RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
@@ -8,38 +8,38 @@ install(TARGETS
     PUBLIC_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
 )
 
-export(EXPORT commonExport
-    NAMESPACE common::  # to simulate a different name and see it works
-    FILE "${CMAKE_CURRENT_BINARY_DIR}/commonTargets.cmake"
+export(EXPORT ps-common-lib-export
+    NAMESPACE ps-common-lib::  # to simulate a different name and see it works
+    FILE "${CMAKE_CURRENT_BINARY_DIR}/ps-common-lib-targets.cmake"
 )
 
-install(EXPORT commonExport 
-    FILE commonTargets.cmake
-    NAMESPACE common::
-    DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/common
+install(EXPORT ps-common-lib-export 
+    FILE ps-common-lib-targets.cmake
+    NAMESPACE ps-common-lib::
+    DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/ps-common-lib
 )
 
 include(CMakePackageConfigHelpers)
 
 write_basic_package_version_file(
-    "${CMAKE_CURRENT_BINARY_DIR}/common-config-version.cmake"
+    "${CMAKE_CURRENT_BINARY_DIR}/ps-common-lib-config-version.cmake"
     VERSION ${PROJECT_VERSION}
     COMPATIBILITY AnyNewerVersion
 )
 
 configure_package_config_file(
-    ${COMMON_ROOT}/cmake/common-config.cmake.in
-    "${CMAKE_CURRENT_BINARY_DIR}/common-config.cmake"
-    INSTALL_DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/common
+    ${COMMON_ROOT}/cmake/ps-common-lib-config.cmake.in
+    "${CMAKE_CURRENT_BINARY_DIR}/ps-common-lib-config.cmake"
+    INSTALL_DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/ps-common-lib
 )
 
 install(FILES
-    "${CMAKE_CURRENT_BINARY_DIR}/common-config.cmake"
-    "${CMAKE_CURRENT_BINARY_DIR}/common-config-version.cmake"
-    DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/common
+    "${CMAKE_CURRENT_BINARY_DIR}/ps-common-lib-config.cmake"
+    "${CMAKE_CURRENT_BINARY_DIR}/ps-common-lib-config-version.cmake"
+    DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/ps-common-lib
 )
 
-set(CPACK_PACKAGE_NAME                  common)
+set(CPACK_PACKAGE_NAME                  ps-common-lib)
 set(CPACK_PACKAGE_VENDOR                "OSTIS AI")
 set(CPACK_PACKAGE_DESCRIPTION_SUMMARY   "Library of common classes and functions")
 set(CPACK_PACKAGE_INSTALL_DIRECTORY     ${CPACK_PACKAGE_NAME})
