+++ ps-common-lib/cmake/ps-common-lib-config.cmake.in
@@ -1,0 +1,3 @@
+@PACKAGE_INIT@
+
+include("${CMAKE_CURRENT_LIST_DIR}/ps-common-lib-targets.cmake")
