+++ ps-common-lib/scripts/clang/check_formatting.sh
