+++ ps-common-lib/scripts/clang/format_code.sh
